Create frozenSet.py
Frozen Sets support only methods and operators that deliver results without altering the frozen set or sets to which they are applied. Frozen sets are immutable objects. A set's components can be changed at any moment, but the components of a frozen set don't change after they've been created.
It returns an empty frozen set if no arguments are supplied.